TestCreateWholeImageFromCBR
