+++ acceptance/openstack/csbs/v1/common.go
@@ -4,11 +4,7 @@ import (
 	"testing"
 
 	"github.com/opentelekomcloud/gophertelekomcloud"
-	"github.com/opentelekomcloud/gophertelekomcloud/acceptance/openstack"
-	"github.com/opentelekomcloud/gophertelekomcloud/acceptance/openstack/cbr/v3"
 	"github.com/opentelekomcloud/gophertelekomcloud/acceptance/tools"
-	"github.com/opentelekomcloud/gophertelekomcloud/openstack/cbr/v3/checkpoint"
-	"github.com/opentelekomcloud/gophertelekomcloud/openstack/cbr/v3/vaults"
 	"github.com/opentelekomcloud/gophertelekomcloud/openstack/csbs/v1/backup"
 	th "github.com/opentelekomcloud/gophertelekomcloud/testhelper"
 )
@@ -70,55 +66,4 @@ func waitForBackupDeleted(client *golangsdk.ServiceClient, secs int, backupID st
 		}
 		return false, nil
 	})
-}
-
-func CreateCBR(t *testing.T, client *golangsdk.ServiceClient) (*vaults.Vault, vaults.AssociateResourcesOpts, checkpoint.CreateOpts, *checkpoint.Checkpoint) {
-	// Create Vault for further backup
-	opts := vaults.CreateOpts{
-		Billing: &vaults.BillingCreate{
-			ConsistentLevel: "crash_consistent",
-			ObjectType:      "disk",
-			ProtectType:     "backup",
-			Size:            100,
-		},
-		Description: "gophertelemocloud testing vault",
-		Name:        tools.RandomString("cbr-test-", 5),
-		Resources:   []vaults.ResourceCreate{},
-	}
-	vault, err := vaults.Create(client, opts)
-	th.AssertNoErr(t, err)
-	t.Cleanup(func() {
-		th.AssertNoErr(t, vaults.Delete(client, vault.ID))
-	})
-
-	// Create Volume
-	volume := openstack.CreateVolume(t)
-	t.Cleanup(func() {
-		openstack.DeleteVolume(t, volume.ID)
-	})
-
-	// Associate server to the vault
-	aOpts := vaults.AssociateResourcesOpts{
-		Resources: []vaults.ResourceCreate{
-			{
-				ID:   volume.ID,
-				Type: "OS::Cinder::Volume",
-			},
-		},
-	}
-	associated, err := vaults.AssociateResources(client, vault.ID, aOpts)
-	th.AssertNoErr(t, err)
-	th.AssertEquals(t, 1, len(associated))
-
-	// Create vault checkpoint
-	optsVault := checkpoint.CreateOpts{
-		VaultID: vault.ID,
-		Parameters: checkpoint.CheckpointParam{
-			Description: "go created backup",
-			Incremental: true,
-			Name:        tools.RandomString("go-checkpoint", 5),
-		},
-	}
-	checkp := v3.CreateCheckpoint(t, client, optsVault)
-	return vault, aOpts, optsVault, checkp
 }
